Olaf's changes
A suggestion on the text below:

OLD:
ISOC maintains liaison relationships and memberships in other
Standards Development Organizations (SDOs), which directly benefits
the IETF.


NEW:
ISOC maintains liaison relationships and memberships in other
Standards Development Organizations (SDOs) and related organizations, 
which directly benefits the IETF.

MOTIVATION:
Introducing some vagueness with ‘related organizations’ since not all of these groups are formal SDOs. For instance, the representation for the IETF in the European Multi Stakeholder Platform on ICT Standardisation is on the ISOC name-plate (although I might be as confused about this as anybody else). Coincidentally the two people representing the IETF there are ISOC staff members - but that is coincidental, they are IAB appointed volunteers.


Which brings me to my next point. ISOC staff does participate in the standards process, they do so as individual contributors. I think it is important to clarify that in section 4 of the the 00 draft. A suggested phrase: Involvement of ISOC’s employees in the standard process e.g. as document editors, or in leadership positions, is as individual contributors rather than on institutional grounds.


Finally, I think it would be nice that if in the ‘ISOC also supports the IETF standards process more indirectly’ we add some verbiage about ISOC in its technical work using the standards developed in the IETF as its basis. An example of that is our Deploy360 program, although that does not necessarily need to be mentioned.


—Olaf (as an individual contributor ;-) )